First pass of the database adapter class
Had to modify a bunch of other classes to get this working properly. The database adapter class is definitely not done yet, and needs to be refactored a bit for better performance, but the implemented methods do work. Login screen now checks if username has been taken before letting users sign up.