Finally working properly
Habits and habit events are no longer in the user object because it made working with firebase an absolute nightmare. I unironically spent 10 hours trying to get it to work with the current structure, but I couldn't get it running properly. To get it working, I restructured how we store data on the app side. Habits are now stored in a HabitList object, and habit events are stored in a HabitEvent object.