Database adapter (#83)
* initial refactor of existing methods

* pushing habits properly.... maybe

* Working more better now.

* I've given in to google

I give in. Changing things to work with how firestore wants them to be

* Finally working properly

Habits and habit events are no longer in the user object because it made working with firebase an absolute nightmare. I unironically spent 10 hours trying to get it to work with the current structure, but I couldn't get it running properly. To get it working, I restructured how we store data on the app side. Habits are now stored in a HabitList object, and habit events are stored in a HabitEvent object.

* add support for habit events

This should hopefully work with habit events when they are added. Hopefully.

* Remove user dependency when adding a habit event

* Add comment explaining -1

Co-authored-by: Christian Jans <cjjans@ualberta.ca>