Add friend requests (#129)
* making progress. Now we test

* We can now send/receive/accept friend requests!

This is probably the hardest code I've had to write in well over a year. It's probably easier to get away with murder than it is to get multiple firebase auth sessions working simultaneously. Even switching between profiles by signing out and signing in the new user is situationally problematic for some reason. So instead we have convoluted tests... because Google. The tests pass tho, and I was able to verify that the right stuff is happening in the firebase console. As an aside, I wonder how long a commit message can be. Is there even a limit?

* add functionality to remove friends

* UI now allows for sending/receiving friend requests and viewing friends!

* fixed bugs

Can no longer friend yourself, or add other people multiple times. Also disabled the user unit tests because a user now needs an instance of the database adapter, which means it cannot be tested as a unit test. Thanks google.

* fix spelling mistakes